Use safe_sparse_dot.
